requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
trl>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
kaggle-environments>=1.12.0
wandb>=0.16.0
evalplus>=0.2.0
pytest>=7.4.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0